// src/types/RegisterUserRequest.ts
export interface RegisterUserRequest {
  idCard?: string
  laserNo?: string
  name?: string
  surname?: string
  birthDate?: string
  username?: string
  password?: string
  mobileNo?: string
  email?: string
  imageIDCardFront?: string
  imageIDCardBack?: string
  imageSelfie?: string
  contentTypeIDCard?: string
  contentTypeSelfie?: string
}